feature: Added support for JSON output
This commit consists of major refactors also that was required to add
JSON output feature. The `Result` and `VerboseResult` classes have been
removed and instead new class are added that seperate the problem of
accumulating the output and actually printing the output, thus following
the first pillar of SOLID principle. There are some TODOs left which
will be implemented in future commits.